finished useTimelineTracks to make it lighter :)
